requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
crontab==0.22.6
Django==3.0.5
django-cors-headers==3.2.1
django-crontab==0.7.1
django-js-asset==1.2.2
django-mptt==0.11.0
mysqlclient==1.4.6
Pillow==7.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
